Batch effect
Simulates gene expression with a known batch effect and "treatment" by adding these effects to a real gene expression dataset. The "treatment" is added to 500 genes. Combat, linear models and linear mixed models methodologies are applied in order to assess the number of true or false positives that each method identifies.